logger: make buffering optional
Issue: KBFS-3986
